feat: set up environment
