+++ index.ts
@@ -1,5 +1,5 @@
 import Jimp from 'jimp';
-import {httpServer} from './src/http_server/index.js';
+import { httpServer } from './src/http_server/index';
 import robot from 'robotjs';
 import { WebSocketServer } from 'ws';
