+++ package.json
@@ -1,22 +1,32 @@
 {
-    "name": "ws_task",
-    "version": "1.0.0",
-    "description": "WebSocket remote control",
-    "main": "index.js",
-    "type": "module",
-    "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js"
-    },
-    "author": "Maria Serykh",
-    "keywords": [],
-    "license": "ISC",
-    "devDependencies": {
-        "nodemon": "^2.0.16"
-    },
-    "dependencies": {
-        "jimp": "^0.16.1",
-        "robotjs": "^0.6.0",
-        "ws": "^8.8.0"
-    }
+  "name": "ws_task",
+  "version": "1.0.0",
+  "description": "WebSocket remote control",
+  "main": "index.ts",
+  "type": "commonjs",
+  "scripts": {
+    "start": "tsc && node ./dist/index.js",
+    "start:dev": "nodemon ./index.ts"
+  },
+  "author": "Maria Serykh",
+  "keywords": [],
+  "license": "ISC",
+  "devDependencies": {
+    "@types/jimp": "^0.2.28",
+    "@types/ws": "^8.5.3",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "eslint": "^8.18.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "nodemon": "^2.0.18",
+    "prettier": "^2.7.1",
+    "typescript": "^4.7.4"
+  },
+  "dependencies": {
+    "jimp": "^0.16.1",
+    "robotjs": "^0.6.0",
+    "ts-node": "^10.8.1",
+    "ws": "^8.8.0"
+  }
 }
